Changed XMSSMTSignature Index to long, also changed the indexTree in verifyMT() to long, changed the calculation of indexTree and indexLeaf to a bit shifting(like Andreas Huelsing in his C-Implementation). There seems to be an issue with the integer cast of indexLeaf. Added tests for manipulated Signatures in MT.

+++ core/src/main/java/org/bouncycastle/pqc/crypto/xmss/XMSSAddress.java
@@ -63,7 +63,7 @@ public abstract class XMSSAddress {
 		this.type = type;
 	}
 
-	public int getKeyAndMask() {
+	public long getKeyAndMask() {
 		return keyAndMask;
 	}
